Move kernelfuture to a separate file
